split the app.js file into smaller files
