+++ src/client/index.js
@@ -1,15 +1,21 @@
 import { processInfo } from "./js/app.js";
 import { calcDate } from "./js/calculateDate.js";
-import { getCurrentWeather, getFutureWeather } from "./js/weatherApiCalls";
+import { getCurrentWeather, getFutureWeather } from "./js/weatherApiCalls.js";
+import { postData } from "./js/postData.js";
+import { getCityCoords } from "./js/getCoords.js";
 
-// console.log(testFuntion)
-// alert('HELLO WORLD')
 document.getElementById("submit").addEventListener("click", processInfo);
 
 
 import './styles/main.scss'
 import'./styles/fonts.scss'
 
-// document.getElementById("submit").addEventListener("click", processInfo);
 
-export { processInfo, calcDate, getFutureWeather, getCurrentWeather}
+export {
+  processInfo,
+  calcDate,
+  getFutureWeather,
+  getCurrentWeather,
+  postData,
+  getCityCoords
+};
