+++ src/client/js/postData.js
@@ -1,0 +1,21 @@
+const postData = async (url = "", data = {}) => {
+  // console.log("POST");
+  const response = await fetch(url, {
+    method: "POST",
+    credentials: "same-origin",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(data),
+  });
+
+  try {
+    const newData = await response.json();
+    console.log(newData);
+    return newData;
+  } catch (error) {
+    console.log("error", error);
+  }
+};
+
+export {postData}
